Implemented Powder Snow Move
Replaced the (will be) unused Vicegrip with Powder Snow.

Removed all references to Vicegrip in all code.

Not sure if removing Powder Snow from the grammar file will break other moves, so left it in for the time being. Will need testing. Also unsure what that even does currently.

Gave Powder Snow the same animation as Ice Punch, but gave it a randomly generated sfx. Playing around with this is beyond my level of caring currently.