Replaced TM20's Rage with Powder Snow
Because almost every Pokemon learns Rage in Gen I, I had to manually go in and edit nearly every single Pokemon's file to either replace it with Powder Snow, or remove Rage entirely.

Changed it so Diana now gives the Powder Snow TM.

Changed the previous location of TM 20 on Route 15 to now be Bide instead.