Fixed Effectiveness Text Glitch
Due to a programming error, in gen 1 the text and sound that are displayed in type match ups is only affected by the secondary type of the defending Pokemon.

This bug doesn't actually affect the damage equation, just the text.

This fix makes sure that the correct text and sound is displayed due to the true results of an attack.

This code is taken directly from the tutorial on fixing the bug.