Fixed Celadon Raichu Sprite
Due to there being no Pokemon with the "Monster" icon evolving via stone, I had to change the Poliwrath into a Raichu with the "Fairy" icon.

Unfortunately, it looks like only 11 different sprites can be loaded per zone. I replaced Monster with Fairy to accommodate the change. Just something to keep in mind for the future.