Modified Celadon Trainer Move Sets
Gave a bunch of Pokemon in the Celadon gym Confuse Ray.

Due to the lack of appropriate ghost Pokemon for the fourth gym, I tried to add more Pokemon that had ghost moves. This involved adding Confuse Ray to a few mons to increase gym diversity and add an appropriate "feel" to the gym.