Added Rare Candy to Safari Zone
The Rare Candy in the Warden's House will be replaced with a Claw Fossil, so I'm putting the Rare Candy in the Safari Zone instead.

Note: This was a PITA to implement. Items have to be grouped together even within a single zone. Their ID number cannot come after signs/NPC dialog.

There is a small chance this could negatively impact some other items/static encounters in zones that come later in the game. If so, this commit could be the culprit.