FeMC Bank1E Fix
Removed Pokemon Blue builds from the makefile.

Moved the slots animations out of the battle animations file and into its own, then added it to the Text 11 ROM that had plenty of space in order to make enough room for the mod to compile.

Further testing needed to ensure that everything is working as it should.