Static Damage and Transform Fixes
Static Damage moves now obey type immunities.

Transform will now turn Transform into Struggle if someone transforms into a mon that knows Transform. As Ditto is not in this game, this should only be encountered when fighting against a Mew with either a Mew, or a Pokemon that knows Metronome.