Replaced Bide with Twister
Added Twister to the game and removed Bide. This required commenting out a large chunk of the battle logic to account for no more Bide.

Also increased the potency of Wing Attack from 40 to 50, and re-arranged moves on some Pokemon to account for Twister, the removal of Bide, and the Potency increase on Wing Attack.